Refactor esp program management into a separate class
